+++ src/main/java/com/zoarial/iot/ServerServer.java
@@ -23,7 +23,7 @@ import java.util.stream.Stream;
 public class ServerServer extends PrintBaseClass implements Runnable {
 
     final private String _hostname;
-    final private int _nodeType;
+    final private byte _nodeType;
     final private boolean _isVolatile;
     final private int _serverPort;
     final private boolean _isHeadCapable;
@@ -54,7 +54,7 @@ public class ServerServer extends PrintBaseClass implements Runnable {
     DatagramSocketHelper _datagramSocketHelper;
 
 
-    public ServerServer(String hostname, UUID uuid, int nodeType, Boolean isVolatile, int serverPort, int messageTimeout, int pingTimeout, boolean isHeadCapable) throws Exception {
+    public ServerServer(String hostname, UUID uuid, byte nodeType, Boolean isVolatile, int serverPort, int messageTimeout, int pingTimeout, boolean isHeadCapable) throws Exception {
         super("ServerServer");
 
         println("Initializing...");
@@ -350,7 +350,7 @@ public class ServerServer extends PrintBaseClass implements Runnable {
         return _serverPort;
     }
 
-    public int getNodeType() {
+    public byte getNodeType() {
         return _nodeType;
     }
